code/requirements.txt
functions-framework==3.*
discord.py
requests
